Starting files
Single-player and 2-player versions, both of which randomize a single square in a 5x5 board to be the battleship. 

Next steps: I should first expand to allow multiple ships and create the new win condition, and also let players in 2-player place their own ships on the board for their opponent to find.